Extract data_loader module from the monolith (split PR 7/N)
Moves the experimental data download + species-specific loading
functions into tfbs_footprinter3/data_loader.py:

- experimentalDataUpdater: first-run + --update base data.tar.gz download
- experimentaldata: per-species tarball download
- experimentalDataUpdater_beta: alternative updater (preserved as-is)
- species_specific_data: load PWM thresholds, GERP locations, CAGE,
  GTEx, GTRD metaclusters, ATAC-Seq, CpG weights, and the CAS p-values
  table for a given species + chromosome

script_dir is redefined at the top of data_loader.py (os.path.dirname
of the module's __file__). Since both modules sit in the package
directory the resolved path is identical.

Behavior-neutral dead-code cleanup: drops ~7 blocks of ##-commented
earlier versions that littered species_specific_data (GERP weights
dict, CAGE keys dict, jasparTFs transcripts dict, ATAC-Seq Encode
variant, etc.). All replaced by their active successors already.

Monolith: 1201 -> 853 lines (-348). 21 tests green, ruff clean,
smoke verifies experimentalDataUpdater / experimentaldata /
species_specific_data all resolve via the tfbs_footprinter3
re-export path.
